+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 - New #57: Add `addStickerToSet`, `deleteStickerFromSet` and `replaceStickerInSet` methods.
 - New #58: Add `setStickerPositionInSet`, `setStickerEmojiList`, `setStickerKeywords` and `setStickerMaskPosition`
   methods.
+- New #60: Add `setStickerSetTitle` method.
 - Chg #24: Move update methods to `Vjik\TelegramBot\Api\Method\Update` namespace, and update types to
   `Vjik\TelegramBot\Api\Type\Update` namespace.
 - Chg #30: Remove `TelegramRequestWithFilesInterface`.